Move RawGridContainer to own file
